缩减plugin微服务的子模块数量 #9231 移除无用的codecc旧接口,改common-codecc为依赖
