+++ src/backend/ci/core/common/common-codecc/src/main/kotlin/com/tencent/devops/plugin/codecc/pojo/CodeccMeasureInfo.kt
