--- src/backend/ci/core/plugin/codecc-plugin/biz-codecc/src/main/kotlin/com/tencent/devops/plugin/codecc/dao/PluginCodeccDao.kt
@@ -1,98 +1,0 @@
-/*
- * Tencent is pleased to support the open source community by making BK-CI 蓝鲸持续集成平台 available.
- *
- * Copyright (C) 2019 THL A29 Limited, a Tencent company.  All rights reserved.
- *
- * BK-CI 蓝鲸持续集成平台 is licensed under the MIT license.
- *
- * A copy of the MIT License is included in this file.
- *
- *
- * Terms of the MIT License:
- * ---------------------------------------------------
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
- * documentation files (the "Software"), to deal in the Software without restriction, including without limitation the
- * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of
- * the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT
- * LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN
- * NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
- * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
- * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package com.tencent.devops.plugin.codecc.dao
-
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.tencent.devops.model.plugin.tables.TPluginCodecc
-import com.tencent.devops.model.plugin.tables.records.TPluginCodeccRecord
-import com.tencent.devops.plugin.codecc.pojo.CodeccCallback
-import org.jooq.DSLContext
-import org.jooq.Result
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.stereotype.Repository
-
-@Repository
-class PluginCodeccDao @Autowired constructor(
-    private val objectMapper: ObjectMapper
-) {
-    fun saveCallback(dslContext: DSLContext, callback: CodeccCallback): Int {
-        with(TPluginCodecc.T_PLUGIN_CODECC) {
-            return dslContext.insertInto(
-                this,
-                PROJECT_ID,
-                PIPELINE_ID,
-                BUILD_ID,
-                TASK_ID,
-                TOOL_SNAPSHOT_LIST
-            )
-                .values(
-                    callback.projectId,
-                    callback.pipelineId,
-                    callback.buildId,
-                    callback.taskId,
-                    objectMapper.writeValueAsString(callback.toolSnapshotList)
-                )
-                .onDuplicateKeyUpdate()
-                .set(TASK_ID, callback.taskId)
-                .set(TOOL_SNAPSHOT_LIST, objectMapper.writeValueAsString(callback.toolSnapshotList))
-                .execute()
-        }
-    }
-
-    fun getCallback(dslContext: DSLContext, buildId: String): TPluginCodeccRecord? {
-        with(TPluginCodecc.T_PLUGIN_CODECC) {
-            return dslContext.selectFrom(this)
-                .where(BUILD_ID.eq(buildId))
-                .fetchOne()
-        }
-    }
-
-    fun getCallbackByProject(dslContext: DSLContext, projectIds: Set<String>): Result<TPluginCodeccRecord>? {
-        with(TPluginCodecc.T_PLUGIN_CODECC) {
-            return dslContext.selectFrom(this)
-                .where(PROJECT_ID.`in`(projectIds))
-                .fetch()
-        }
-    }
-
-    fun getCallbackByPipeline(dslContext: DSLContext, pipelineIds: Set<String>): Result<TPluginCodeccRecord>? {
-        with(TPluginCodecc.T_PLUGIN_CODECC) {
-            return dslContext.selectFrom(this)
-                .where(PIPELINE_ID.`in`(pipelineIds))
-                .fetch()
-        }
-    }
-
-    fun getCallbackByBuildId(dslContext: DSLContext, buildIds: Set<String>): Result<TPluginCodeccRecord> {
-        with(TPluginCodecc.T_PLUGIN_CODECC) {
-            return dslContext.selectFrom(this)
-                .where(BUILD_ID.`in`(buildIds))
-                .fetch()
-        }
-    }
-}
