+++ src/backend/ci/core/quality/biz-quality/build.gradle.kts
@@ -37,8 +37,7 @@ dependencies {
     api(project(":core:common:common-archive"))
     api(project(":core:common:common-auth:common-auth-rbac"))
 
-    api(project(":core:plugin:codecc-plugin:common-codecc"))
-    api(project(":core:plugin:codecc-plugin:api-codecc"))
+    api(project(":core:common:common-codecc"))
 
     api(project(":core:project:api-project"))
     api(project(":core:notify:api-notify"))
