+++ src/backend/ci/core/stream/biz-stream/build.gradle.kts
@@ -43,7 +43,7 @@ dependencies {
     api(project(":core:plugin:api-plugin"))
     api(project(":core:quality:api-quality"))
     api(project(":core:auth:api-auth"))
-    api(project(":core:plugin:codecc-plugin:api-codecc"))
+    api(project(":core:common:common-codecc"))
     api(project(":core:common:common-pipeline-yaml"))
     api(project(":core:notify:api-notify"))
