+++ src/backend/ci/settings.gradle.kts
@@ -81,7 +81,7 @@ include(":core:common:common-auth:common-auth-v3")
 include(":core:common:common-wechatwork")
 include(":core:common:common-auth:common-auth-rbac")
 include(":core:common:common-pipeline-yaml")
-
+include(":core:common:common-codecc")
 include(":core:common:common-notify")
 include(":core:common:common-websocket")
 include(":core:common:common-util")
@@ -238,11 +238,6 @@ include(":core:plugin:api-plugin")
 include(":core:plugin:biz-plugin")
 include(":core:plugin:boot-plugin")
 include(":core:plugin:model-plugin")
-
-include(":core:plugin:codecc-plugin")
-include(":core:plugin:codecc-plugin:common-codecc")
-include(":core:plugin:codecc-plugin:api-codecc")
-include(":core:plugin:codecc-plugin:biz-codecc")
 
 include(":core:worker")
 include(":core:worker:worker-agent")
